refactor/ some cleanup

+++ app/src/main/java/com/kidrich/turbotrack/VerticalBarChartFragment.kt
@@ -1,5 +1,6 @@
 package com.kidrich.turbotrack
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -220,18 +221,24 @@ private class BarChartOnChartValueSelectedListener : OnChartValueSelectedListene
     }
 
     private fun onDetailButtonClicked(ingredients: List<Ingredient>) {
-        // Add your logic for handling button click with specific ingredients
+        // Add logic for handling button click with specific ingredients
     }
 
     private fun onRemoveButtonClicked(meal: Meal, mealViewModel: MealViewModel, view: LinearLayout, component: View) {
-        CoroutineScope(Dispatchers.Main).launch {
-            view.removeView(component)
-            mealViewModel.onEvent(MealEvent.DeleteMeal(meal))
-        }
+            AlertDialog.Builder(this.activity)
+                .setIcon(com.kidrich.turbotrack.R.drawable.baseline_delete_forever_24)
+                .setTitle("Are you sure you want to delete this Meal?")
+                .setPositiveButton("Alrighty") {_, _ ->
+                    CoroutineScope(Dispatchers.Main).launch {
+                        view.removeView(component)
+                        mealViewModel.onEvent(MealEvent.DeleteMeal(meal))
+                    }
+                }.create().show()
     }
 
     override fun onNothingSelected() {
         val informationLayout = activity.findViewById<LinearLayout>(R.id.meal_clicked_information)
         informationLayout.removeAllViews()
     }
+
 }
